requirements.txt
Django==2.0.7
djangorestframework==3.9.1
djangorestframework-camel-case==1.0.0
psycopg2==2.7.5
pytz==2018.5
six==1.12.0